% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Permutate.R
\name{add.col.self}
\alias{add.col.self}
\title{Add column without restriction
Add new column to the new dataframe without restriction}
\usage{
add.col.self(dataframe, new.vector)
}
\arguments{
\item{dataframe}{The data.frame with unrestricted shape, including the empty dataframe}

\item{new.vector}{The vector whith unrestricted length}
}
\value{
A new matrix
}
\description{
Add column without restriction
Add new column to the new dataframe without restriction
}
