% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Objects.R
\docType{class}
\name{Assays-class}
\alias{Assays-class}
\alias{Assays}
\title{The Assays Class}
\description{
The Assays Class is an intermediate data storage class that stores the raw normalized data
and the normalized data filtered by Intercellular Interaction Database.
}
\section{Slots}{

\describe{
\item{\code{RNData}}{An original normalized data}

\item{\code{FNData}}{An filtered normalized data}
}}

