% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generics.R, R/Permutate.R
\name{ccptp_ptest}
\alias{ccptp_ptest}
\alias{ccptp_ptest.default}
\alias{ccptp_ptest.FindPairs}
\title{Cell-Cell interaction per Time-Point Permutate Test}
\usage{
ccptp_ptest(object, ...)

\method{ccptp_ptest}{default}(object, ...)

\method{ccptp_ptest}{FindPairs}(object, number, i, Asample_size,
  Bsample_size, seed, ...)
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}

\item{number}{Permutation numbers}

\item{i}{The days of intercellular interaction, such as "day1"}

\item{Asample_size}{The number of cell expressing symbol A virtually. We generalized the concept of permutation test.
In the traditional permutation test, the number of random substitutions should be the total number of
corresponding targets. But, in our permutation test, we set this number as an adjustable parameter. This paramter
have two funcitons: one is to provide a fixed background distribution, and other is to reduce the impact of low number
cells in the permutation.}

\item{Bsample_size}{The number of cell expressing symbol B virtually.}

\item{seed}{random seed}
}
\description{
Permutate Test for different cell-cell interaction of one time point
}
