% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Objects.R
\docType{class}
\name{Output-class}
\alias{Output-class}
\alias{Output}
\title{The Output object
The Output class store the caculated result}
\description{
The Output object
The Output class store the caculated result
}
\section{Slots}{

\describe{
\item{\code{output.pvalue}}{output the pvalue result}

\item{\code{output.strength}}{output the interactive strength result}
}}

