% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Objects.R
\name{BuildFP}
\alias{BuildFP}
\title{Setup the FindPairs object}
\usage{
BuildFP(data, meta.data, interactive.type, ident = c("mtp", "stp"))
}
\arguments{
\item{data}{the normalized data of gene expression}

\item{meta.data}{annotation of cells. "Cell types" must be annotated. If it is multi-times-point, the times
information should be added to the dataframe.}

\item{interactive.type}{The \strong{feature} expressed by \strong{cell types}. The dataframe must contain two variables.
This parameter needs to be set manually, and it must follow the  format below.
\itemize{
  \item the \strong{cell types} must be in the first column
  \item the \strong{gene feature} must be in the second column, and \strong{gene feature} have two varibales:
  \strong{"SYMBOLA"} means this gene feature is "SYMBOLA" feature and \strong{"SYMBOLB"} means this gene feature is
  "SYMBOLB" feature.
}}

\item{ident}{Determine whether the input data contains multi-times information. There are two arguments, mtp:
multi-times-point or stp: single-time-point.}
}
\description{
Setup the FindPairs object
}
