% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generics.R, R/Permutate.R
\name{Calculate_score}
\alias{Calculate_score}
\alias{Calculate_score.default}
\alias{Calculate_score.FindPairs}
\title{Calculate the interacellular interacion score}
\usage{
Calculate_score(object, ...)

\method{Calculate_score}{default}(object, ...)

\method{Calculate_score}{FindPairs}(object, metric = c("mean", "ratio"),
  ...)
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}

\item{metric}{A metric of interaction. Set to two methods: ratio and mean.
\itemize{
  \item ratio: The percentage of cells with gene expression > 0
  \item mean: The mean of gene expression
}}
}
\value{
Returns FindPairs object with the score calculation stored in the Output slot
}
\description{
Calculate the cell-cell interaction score.
}
