% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Objects.R
\docType{class}
\name{FindPairs-class}
\alias{FindPairs-class}
\alias{FindPairs}
\title{The FindPairs Class}
\description{
The FindPairs Class is a repesentation of single-cell expression data for R.
This Clsss is set to better calculate the intercellular intercation.
}
\section{Slots}{

\describe{
\item{\code{Assays}}{A list of assays of this project}

\item{\code{MData}}{Contains meta-information about each cells}

\item{\code{IA_Type}}{User-defined intercellular interaction need to be calculated}

\item{\code{P}}{User-defined parameters}

\item{\code{PPi}}{Curated Protein-Protein Interaciton Database}

\item{\code{Output}}{A list of Output of this project}
}}

