% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generics.R, R/Permutate.R
\name{FP_test}
\alias{FP_test}
\alias{FP_test.default}
\alias{FP_test.FindPairs}
\title{FindPairs object permutate test
Permutate Test on all intercellular interaction}
\usage{
FP_test(object, ...)

\method{FP_test}{default}(object, ...)

\method{FP_test}{FindPairs}(object, Asample_size = 3000,
  Bsample_size = 3000, target = c("STP", "CCpTP", "TPpCC"),
  fdr = 0.01, fdr_method = "BH", permutation_number = 5000,
  dopar = FALSE, dopar.numbers = 2, random.seed = 42, ...)
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}

\item{Asample_size}{Permutating the number of cell expressing symbol A. Default 200.We generalized the concept of
permutation test. In the traditional permutation test, the number of random substitutions should be the total number
of corresponding targets(the target cell type). But, in our permutation test, we set this number as an adjustable
parameter. This paramterhave two funcitons: one is to provide a fixed background distribution, and other is to
reduce the impact of low number cells in the permutation.}

\item{Bsample_size}{Permutating the number of cell expressing symbol B. Default 200'}

\item{target}{This agrument is corresponding to the agrument \code{target} of function \code{Calculate_score}.If
the data is "stp", setting \code{target = "STP"}, this will do the permutation test under this condition. If
this is the "mtp" data, the "target" has two potential parameters. 1. "CCpTP", \strong{cell-cell per time point}:
this will do the permutation test for the interaciton of different cell types at time poit. 2. "TPpCC",
\strong{time point per cell-cell}: this will do the permutation test for the cell-cell interaction of different
time points for each cell-cell interaciton.}

\item{fdr}{Filter for significant pvalue. Default 0.01}

\item{fdr_method}{Corrected method in multiple hypothesis testing. Default "BH".}

\item{permutation_number}{The number of permutation. Default 5000}

\item{dopar}{Whether Parallel computation is allowed. Default FLASE.}

\item{dopar.numbers}{If the \code{do.par = TRUE}, multiple cores will be used for parallel operations. Default 2.}

\item{random.seed}{Set the random seed. Default 42}
}
\value{
Returns FindPairs object with the pvalue calculation stored in the Output slot


}
\description{
FindPairs object permutate test
Permutate Test on all intercellular interaction
}
